Move progress bar logic
